+++ lib/resource/template/progressView/ProgressSpinkit.dart
@@ -6,17 +6,18 @@ import 'package:flutter/material.dart';
 
 class ProgressSpinkit {
 
+
   static Widget centerPage(BuildContext context ) {
     var stack = Stack(children: [
       EmptyView.empty(   DeviceTools.getWidth(context), DeviceTools.getHalfHeight(context)  ),
-      Positioned(child:  ProgressSpinkit.get(context), left: 0, right: 0, top: DeviceTools.getHalfHeight(context)-50 )
+      Positioned(child:  ProgressSpinkit.get( ), left: 0, right: 0, top: DeviceTools.getHalfHeight(context)-50 )
     ],);
 
     return stack;
   }
 
 
-  static Widget get(BuildContext context ) {
+  static Widget get(  ) {
 
     HexColor c1 =  DSColor.ds_progress_fade_1;
     HexColor c2 =  DSColor.ds_progress_fade_2 ;
@@ -34,8 +35,8 @@ class ProgressSpinkit {
     return spinkit;
   }
 
-  static Widget getFixedSize(BuildContext context , double size){
-    var prg = ProgressSpinkit.get(  context);
+  static Widget getFixedSize(   double size){
+    var prg = ProgressSpinkit.get(   );
 
     var sizeBox = SizedBox( child: prg, width: size, height:  size ,);